test1

+++ portal-impl/src/com/liferay/portal/repository/liferayrepository/LiferayLocalRepository.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -15,41 +15,40 @@
 package com.liferay.portal.repository.liferayrepository;
 
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.repository.LocalRepository;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.repository.model.FileVersion;
 import com.liferay.portal.kernel.repository.model.Folder;
+import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.SortedArrayList;
 import com.liferay.portal.kernel.util.UnicodeProperties;
-import com.liferay.portal.model.Repository;
 import com.liferay.portal.repository.liferayrepository.model.LiferayFileEntry;
 import com.liferay.portal.repository.liferayrepository.model.LiferayFileVersion;
 import com.liferay.portal.repository.liferayrepository.model.LiferayFolder;
 import com.liferay.portal.service.RepositoryLocalService;
 import com.liferay.portal.service.RepositoryService;
+import com.liferay.portal.service.ResourceLocalService;
 import com.liferay.portal.service.ServiceContext;
-import com.liferay.portlet.documentlibrary.NoSuchFileEntryException;
-import com.liferay.portlet.documentlibrary.NoSuchFileVersionException;
-import com.liferay.portlet.documentlibrary.NoSuchFolderException;
 import com.liferay.portlet.documentlibrary.model.DLFileEntry;
 import com.liferay.portlet.documentlibrary.model.DLFileVersion;
 import com.liferay.portlet.documentlibrary.model.DLFolder;
 import com.liferay.portlet.documentlibrary.service.DLAppHelperLocalService;
 import com.liferay.portlet.documentlibrary.service.DLFileEntryLocalService;
 import com.liferay.portlet.documentlibrary.service.DLFileEntryService;
+import com.liferay.portlet.documentlibrary.service.DLFileEntryTypeLocalService;
 import com.liferay.portlet.documentlibrary.service.DLFileVersionLocalService;
 import com.liferay.portlet.documentlibrary.service.DLFileVersionService;
 import com.liferay.portlet.documentlibrary.service.DLFolderLocalService;
 import com.liferay.portlet.documentlibrary.service.DLFolderService;
+import com.liferay.portlet.documentlibrary.util.RepositoryModelUtil;
+import com.liferay.portlet.documentlibrary.util.comparator.DLFileEntryOrderByComparator;
 import com.liferay.portlet.dynamicdatamapping.storage.Fields;
 
 import java.io.File;
 import java.io.InputStream;
 
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -64,48 +63,36 @@ public class LiferayLocalRepository
 		DLAppHelperLocalService dlAppHelperLocalService,
 		DLFileEntryLocalService dlFileEntryLocalService,
 		DLFileEntryService dlFileEntryService,
+		DLFileEntryTypeLocalService dlFileEntryTypeLocalService,
 		DLFileVersionLocalService dlFileVersionLocalService,
 		DLFileVersionService dlFileVersionService,
 		DLFolderLocalService dlFolderLocalService,
-		DLFolderService dlFolderService, long repositoryId) {
+		DLFolderService dlFolderService,
+		ResourceLocalService resourceLocalService, long groupId,
+		long repositoryId, long dlFolderId) {
 
 		super(
 			repositoryLocalService, repositoryService, dlAppHelperLocalService,
 			dlFileEntryLocalService, dlFileEntryService,
-			dlFileVersionLocalService, dlFileVersionService,
-			dlFolderLocalService, dlFolderService, repositoryId);
+			dlFileEntryTypeLocalService, dlFileVersionLocalService,
+			dlFileVersionService, dlFolderLocalService, dlFolderService,
+			resourceLocalService, groupId, repositoryId, dlFolderId);
 	}
 
-	public LiferayLocalRepository(
-		RepositoryLocalService repositoryLocalService,
-		RepositoryService repositoryService,
-		DLAppHelperLocalService dlAppHelperLocalService,
-		DLFileEntryLocalService dlFileEntryLocalService,
-		DLFileEntryService dlFileEntryService,
-		DLFileVersionLocalService dlFileVersionLocalService,
-		DLFileVersionService dlFileVersionService,
-		DLFolderLocalService dlFolderLocalService,
-		DLFolderService dlFolderService, long folderId, long fileEntryId,
-		long fileVersionId) {
-
-		super(
-			repositoryLocalService, repositoryService, dlAppHelperLocalService,
-			dlFileEntryLocalService, dlFileEntryService,
-			dlFileVersionLocalService, dlFileVersionService,
-			dlFolderLocalService, dlFolderService, folderId, fileEntryId,
-			fileVersionId);
-	}
-
+	@Override
 	public FileEntry addFileEntry(
 			long userId, long folderId, String sourceFileName, String mimeType,
 			String title, String description, String changeLog, File file,
 			ServiceContext serviceContext)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		long fileEntryTypeId = ParamUtil.getLong(
-			serviceContext, "fileEntryTypeId", -1L);
+			serviceContext, "fileEntryTypeId",
+			getDefaultFileEntryTypeId(serviceContext, folderId));
+
 		Map<String, Fields> fieldsMap = getFieldsMap(
 			serviceContext, fileEntryTypeId);
+
 		long size = 0;
 
 		if (file != null) {
@@ -122,14 +109,17 @@ public class LiferayLocalRepository
 		return new LiferayFileEntry(dlFileEntry);
 	}
 
+	@Override
 	public FileEntry addFileEntry(
 			long userId, long folderId, String sourceFileName, String mimeType,
 			String title, String description, String changeLog, InputStream is,
 			long size, ServiceContext serviceContext)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		long fileEntryTypeId = ParamUtil.getLong(
-			serviceContext, "fileEntryTypeId", -1L);
+			serviceContext, "fileEntryTypeId",
+			getDefaultFileEntryTypeId(serviceContext, folderId));
+
 		Map<String, Fields> fieldsMap = getFieldsMap(
 			serviceContext, fileEntryTypeId);
 
@@ -143,16 +133,18 @@ public class LiferayLocalRepository
 		return new LiferayFileEntry(dlFileEntry);
 	}
 
+	@Override
 	public Folder addFolder(
 			long userId, long parentFolderId, String title, String description,
 			ServiceContext serviceContext)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		boolean mountPoint = ParamUtil.getBoolean(serviceContext, "mountPoint");
 
 		DLFolder dlFolder = dlFolderLocalService.addFolder(
 			userId, getGroupId(), getRepositoryId(), mountPoint,
-			toFolderId(parentFolderId), title, description, serviceContext);
+			toFolderId(parentFolderId), title, description, false,
+			serviceContext);
 
 		return new LiferayFolder(dlFolder);
 	}
@@ -162,33 +154,36 @@ public class LiferayLocalRepository
 		UnicodeProperties typeSettingsProperties) {
 	}
 
-	public void deleteAll() throws PortalException, SystemException {
+	@Override
+	public void deleteAll() throws PortalException {
 		dlFolderLocalService.deleteAll(getGroupId());
 	}
 
-	public void deleteFileEntry(long fileEntryId)
-		throws PortalException, SystemException {
-
+	@Override
+	public void deleteFileEntry(long fileEntryId) throws PortalException {
 		dlFileEntryLocalService.deleteFileEntry(fileEntryId);
 	}
 
-	public void deleteFolder(long folderId)
-		throws PortalException, SystemException {
+	@Override
+	public void deleteFolder(long folderId) throws PortalException {
+		DLFolder dlFolder = dlFolderLocalService.fetchFolder(folderId);
 
-		dlFolderLocalService.deleteFolder(folderId);
+		if (dlFolder != null) {
+			dlFolderLocalService.deleteFolder(folderId);
+		}
 	}
 
-	public FileEntry getFileEntry(long fileEntryId)
-		throws PortalException, SystemException {
-
+	@Override
+	public FileEntry getFileEntry(long fileEntryId) throws PortalException {
 		DLFileEntry dlFileEntry = dlFileEntryLocalService.getFileEntry(
 			fileEntryId);
 
 		return new LiferayFileEntry(dlFileEntry);
 	}
 
+	@Override
 	public FileEntry getFileEntry(long folderId, String title)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		DLFileEntry dlFileEntry = dlFileEntryLocalService.getFileEntry(
 			getGroupId(), toFolderId(folderId), title);
@@ -196,9 +191,8 @@ public class LiferayLocalRepository
 		return new LiferayFileEntry(dlFileEntry);
 	}
 
-	public FileEntry getFileEntryByUuid(String uuid)
-		throws PortalException, SystemException {
-
+	@Override
+	public FileEntry getFileEntryByUuid(String uuid) throws PortalException {
 		DLFileEntry dlFileEntry =
 			dlFileEntryLocalService.getFileEntryByUuidAndGroupId(
 				uuid, getGroupId());
@@ -206,8 +200,9 @@ public class LiferayLocalRepository
 		return new LiferayFileEntry(dlFileEntry);
 	}
 
+	@Override
 	public FileVersion getFileVersion(long fileVersionId)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		DLFileVersion dlFileVersion = dlFileVersionLocalService.getFileVersion(
 			fileVersionId);
@@ -215,16 +210,16 @@ public class LiferayLocalRepository
 		return new LiferayFileVersion(dlFileVersion);
 	}
 
-	public Folder getFolder(long folderId)
-		throws PortalException, SystemException {
-
+	@Override
+	public Folder getFolder(long folderId) throws PortalException {
 		DLFolder dlFolder = dlFolderLocalService.getFolder(folderId);
 
 		return new LiferayFolder(dlFolder);
 	}
 
+	@Override
 	public Folder getFolder(long parentFolderId, String title)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		DLFolder dlFolder = dlFolderLocalService.getFolder(
 			getGroupId(), toFolderId(parentFolderId), title);
@@ -232,10 +227,24 @@ public class LiferayLocalRepository
 		return new LiferayFolder(dlFolder);
 	}
 
+	@Override
+	public List<FileEntry> getRepositoryFileEntries(
+		long rootFolderId, int start, int end,
+		OrderByComparator<FileEntry> obc) {
+
+		List<DLFileEntry> dlFileEntries =
+			dlFileEntryLocalService.getGroupFileEntries(
+				getGroupId(), 0, rootFolderId, start, end,
+				DLFileEntryOrderByComparator.getOrderByComparator(obc));
+
+		return RepositoryModelUtil.toFileEntries(dlFileEntries);
+	}
+
+	@Override
 	public FileEntry moveFileEntry(
 			long userId, long fileEntryId, long newFolderId,
 			ServiceContext serviceContext)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		DLFileEntry dlFileEntry = dlFileEntryLocalService.moveFileEntry(
 			userId, fileEntryId, toFolderId(newFolderId), serviceContext);
@@ -243,27 +252,43 @@ public class LiferayLocalRepository
 		return new LiferayFileEntry(dlFileEntry);
 	}
 
+	@Override
+	public Folder moveFolder(
+			long userId, long folderId, long parentFolderId,
+			ServiceContext serviceContext)
+		throws PortalException {
+
+		DLFolder dlFolder = dlFolderLocalService.moveFolder(
+			userId, folderId, parentFolderId, serviceContext);
+
+		return new LiferayFolder(dlFolder);
+	}
+
+	@Override
 	public void updateAsset(
 			long userId, FileEntry fileEntry, FileVersion fileVersion,
 			long[] assetCategoryIds, String[] assetTagNames,
 			long[] assetLinkEntryIds)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		dlAppHelperLocalService.updateAsset(
 			userId, fileEntry, fileVersion, assetCategoryIds, assetTagNames,
 			assetLinkEntryIds);
 	}
 
+	@Override
 	public FileEntry updateFileEntry(
 			long userId, long fileEntryId, String sourceFileName,
 			String mimeType, String title, String description, String changeLog,
 			boolean majorVersion, File file, ServiceContext serviceContext)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		long fileEntryTypeId = ParamUtil.getLong(
 			serviceContext, "fileEntryTypeId", -1L);
+
 		Map<String, Fields> fieldsMap = getFieldsMap(
 			serviceContext, fileEntryTypeId);
+
 		long size = 0;
 
 		if (file != null) {
@@ -278,15 +303,17 @@ public class LiferayLocalRepository
 		return new LiferayFileEntry(dlFileEntry);
 	}
 
+	@Override
 	public FileEntry updateFileEntry(
 			long userId, long fileEntryId, String sourceFileName,
 			String mimeType, String title, String description, String changeLog,
 			boolean majorVersion, InputStream is, long size,
 			ServiceContext serviceContext)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		long fileEntryTypeId = ParamUtil.getLong(
 			serviceContext, "fileEntryTypeId", -1L);
+
 		Map<String, Fields> fieldsMap = getFieldsMap(
 			serviceContext, fileEntryTypeId);
 
@@ -298,10 +325,11 @@ public class LiferayLocalRepository
 		return new LiferayFileEntry(dlFileEntry);
 	}
 
+	@Override
 	public Folder updateFolder(
 			long folderId, long parentFolderId, String title,
 			String description, ServiceContext serviceContext)
-		throws PortalException, SystemException {
+		throws PortalException {
 
 		long defaultFileEntryTypeId = ParamUtil.getLong(
 			serviceContext, "defaultFileEntryTypeId");
@@ -323,84 +351,5 @@ public class LiferayLocalRepository
 
 		return typeSettingsProperties;
 	}
-
-	@Override
-	protected void initByFileEntryId(long fileEntryId) {
-		try {
-			DLFileEntry dlFileEntry = dlFileEntryLocalService.getFileEntry(
-				fileEntryId);
-
-			initByRepositoryId(dlFileEntry.getRepositoryId());
-		}
-		catch (Exception e) {
-			if (_log.isTraceEnabled()) {
-				if (e instanceof NoSuchFileEntryException) {
-					_log.trace(e.getMessage());
-				}
-				else {
-					_log.trace(e, e);
-				}
-			}
-		}
-	}
-
-	@Override
-	protected void initByFileVersionId(long fileVersionId) {
-		try {
-			DLFileVersion dlFileVersion =
-				dlFileVersionLocalService.getFileVersion(fileVersionId);
-
-			initByRepositoryId(dlFileVersion.getRepositoryId());
-		}
-		catch (Exception e) {
-			if (_log.isTraceEnabled()) {
-				if (e instanceof NoSuchFileVersionException) {
-					_log.trace(e.getMessage());
-				}
-				else {
-					_log.trace(e, e);
-				}
-			}
-		}
-	}
-
-	@Override
-	protected void initByFolderId(long folderId) {
-		try {
-			DLFolder dlFolder = dlFolderLocalService.getFolder(folderId);
-
-			initByRepositoryId(dlFolder.getRepositoryId());
-		}
-		catch (Exception e) {
-			if (_log.isTraceEnabled()) {
-				if (e instanceof NoSuchFolderException) {
-					_log.trace(e.getMessage());
-				}
-				else {
-					_log.trace(e, e);
-				}
-			}
-		}
-	}
-
-	@Override
-	protected void initByRepositoryId(long repositoryId) {
-		setGroupId(repositoryId);
-		setRepositoryId(repositoryId);
-
-		try {
-			Repository repository = repositoryLocalService.getRepository(
-				repositoryId);
-
-			setDlFolderId(repository.getDlFolderId());
-			setGroupId(repository.getGroupId());
-			setRepositoryId(repository.getRepositoryId());
-		}
-		catch (Exception e) {
-		}
-	}
-
-	private static Log _log = LogFactoryUtil.getLog(
-		LiferayLocalRepository.class);
 
 }
