+++ portal-impl/src/com/liferay/portal/util/JarUtil.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -27,6 +27,7 @@ import com.liferay.portal.kernel.util.ProgressStatusConstants;
 import com.liferay.portal.kernel.util.ProgressTracker;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.model.CompanyConstants;
+import com.liferay.portal.spring.context.PortalContextLoaderLifecycleThreadLocal;
 import com.liferay.portlet.documentlibrary.store.DLStoreUtil;
 
 import java.io.File;
@@ -60,7 +61,9 @@ public class JarUtil {
 
 		setProgressStatus(progressTracker, ProgressStatusConstants.COPYING);
 
-		if (PropsValues.CLUSTER_LINK_ENABLED) {
+		if (PropsValues.CLUSTER_LINK_ENABLED &&
+			!PortalContextLoaderLifecycleThreadLocal.isInitializing()) {
+
 			try {
 				DLStoreUtil.deleteFile(
 					_REPOSITORY, _REPOSITORY, _FILE_PATH + name);
@@ -173,6 +176,6 @@ public class JarUtil {
 	private static Log _log = LogFactoryUtil.getLog(JarUtil.class);
 
 	private static MethodKey _installJarKey = new MethodKey(
-		JarUtil.class.getName(), "installJar", boolean.class, String.class);
+		JarUtil.class, "installJar", boolean.class, String.class);
 
 }
